split wrapper tests into their own modules
