+++ symr_tests/symr/test_symformer_wrapper.py
@@ -1,0 +1,71 @@
+
+import os
+
+import unittest
+
+import numpy as np
+import sympy as sp
+
+from symr.wrappers import BaseWrapper
+from symr.nsrts_wrapper import NSRTSWrapper
+from symr.symgpt_wrapper import SymGPTWrapper
+from symr.symformer_wrapper import SymformerWrapper
+
+
+class TestSymformerWrapper(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+    def test_instantiate(self):
+
+        model = SymformerWrapper(use_bfgs=True)
+
+        my_inputs = {"x_1": np.arange(-1.0,1.0,0.035)}
+        y = np.array(my_inputs["x_1"]**2)
+
+        expression, info = model(target=y, **my_inputs)
+
+        sp_expression = sp.sympify(expression)
+        fn_expression = sp.lambdify("x_1", expression)
+        
+        _ = fn_expression(**my_inputs)
+
+        self.assertEqual(str, type(expression))
+
+    def test_call_without_bfgs(self):
+
+        model = SymformerWrapper(use_bfgs=False)
+        my_inputs = {"x_1": np.arange(-1.0,1.0,0.035)}
+        y = np.array(my_inputs["x_1"]**2)
+
+        expression, info = model(target=y, **my_inputs)
+
+        sp_expression = sp.sympify(expression)
+        fn_expression = sp.lambdify("x_1", expression)
+        
+        _ = fn_expression(**my_inputs)
+
+        self.assertEqual(str, type(expression))
+                
+    def test_symformer_multi(self):
+
+        model = SymformerWrapper(use_bfgs=True, model_name="symformer-bivariate")
+        my_inputs = {"x_1": np.arange(-1.0,1.0,0.1),\
+                "x_2": np.arange(-1.0,1.0,0.1),\
+                }
+        y = np.array(my_inputs["x_1"]**2+ np.sin(my_inputs["x_2"]))
+        expression, info = model(target=y, **my_inputs)
+
+        my_vars = ",".join([key for key in my_inputs.keys()])
+        sp_expression = sp.sympify(expression)
+        fn_expression = sp.lambdify(my_vars, expression)
+        
+        _ = fn_expression(**my_inputs)
+
+        self.assertEqual(str, type(expression))
+        self.assertTrue(True)
+
+if __name__ == "__main__": #pragma: no cover    
+
+    unittest.main(verbosity=2)
