+++ symr_tests/symr/test_wrappers.py
@@ -10,8 +10,6 @@ from symr.nsrts_wrapper import NSRTSWrapper
 from symr.symgpt_wrapper import SymGPTWrapper
 from symr.symformer_wrapper import SymformerWrapper
 
-
-
 class TestBaseWrapper(unittest.TestCase):
 
     def setUp(self):
@@ -54,171 +52,6 @@ class TestBaseWrapper(unittest.TestCase):
         self.assertTrue(wrapper())
         self.assertTrue(wrapper.load_parameters())
 
-class TestSymGPTWrapper(unittest.TestCase):
-    
-    def setUp(self):
-        pass
-
-    def test_instantiate(self):
-        
-        model = SymGPTWrapper() 
-
-        self.assertTrue(True)
-
-    def test_call(self):
-
-        model = SymGPTWrapper() 
-
-        my_inputs = {"z": np.arange(-1,1.0,0.01)}
-        y = np.array(my_inputs["z"]**2)
-
-        expression, info = model(target=y, **my_inputs)
-
-        my_vars = ",".join([key for key in my_inputs.keys()])
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify(my_vars, expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-
-    def test_call_w_bfgs(self):
-
-        model = SymGPTWrapper(use_bfgs=True)
-
-        my_inputs = {"x2": np.arange(-1,1.0,0.01)}
-        y = np.array(my_inputs["x2"]**2)
-
-        expression, info = model(target=y, **my_inputs)
-
-        my_vars = ",".join([key for key in my_inputs.keys()])
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify(my_vars, expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-
-class TestNSRTSWrapper(unittest.TestCase):
-
-    def setUp(self):
-        pass
-
-    def test_instantiate(self):
-
-        nsrts = NSRTSWrapper(use_bfgs=True)
-        my_inputs = {"x_1": np.arange(0,1.0,0.1)}
-        y = np.array(my_inputs["x_1"]**2)
-
-        expression, info = nsrts(target=y, **my_inputs)
-
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify("x_1", expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-
-    def test_call_without_bfgs(self):
-
-        nsrts = NSRTSWrapper(use_bfgs=False)
-        my_inputs = {"x_1": np.arange(0,1.0,0.1)}
-        y = np.array(my_inputs["x_1"]**2)
-
-        expression, info = nsrts(target=y, **my_inputs)
-
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify("x_1", expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-                
-    def test_nsrts_multi(self):
-
-        nsrts = NSRTSWrapper(use_bfgs=True)
-        my_inputs = {"x_1": np.arange(0,1.0,0.1),\
-                "x_2": np.arange(0,1.0,0.1),\
-                "x_3": np.arange(0,1.0,0.1),\
-                }
-        y = np.array(my_inputs["x_1"]**2+ np.sin(my_inputs["x_2"]) + my_inputs["x_3"])
-
-        expression, info = nsrts(target=y, **my_inputs)
-
-        my_vars = ",".join([key for key in my_inputs.keys()])
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify(my_vars, expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-        self.assertTrue(True)
-
-        my_inputs = {"x_1": np.arange(0,1.0,0.1),\
-                "x_2": np.arange(0,1.0,0.1),\
-                }
-
-        my_vars = ",".join([key for key in my_inputs.keys()])
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify(my_vars, expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-
-class TestSymformerWrapper(unittest.TestCase):
-
-    def setUp(self):
-        pass
-
-    def test_instantiate(self):
-
-        model = SymformerWrapper(use_bfgs=True)
-
-        my_inputs = {"x_1": np.arange(0,1.0,0.1)}
-        y = np.array(my_inputs["x_1"]**2)
-
-        expression, info = model(target=y, **my_inputs)
-
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify("x_1", expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-
-    def test_call_without_bfgs(self):
-
-        model = SymformerWrapper(use_bfgs=False)
-        my_inputs = {"x_1": np.arange(0,1.0,0.1)}
-        y = np.array(my_inputs["x_1"]**2)
-
-        expression, info = model(target=y, **my_inputs)
-
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify("x_1", expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-                
-    def test_symformer_multi(self):
-
-        model = SymformerWrapper(use_bfgs=True, model_name="symformer-bivariate")
-        my_inputs = {"x_1": np.arange(0,1.0,0.1),\
-                "x_2": np.arange(0,1.0,0.1),\
-                }
-        y = np.array(my_inputs["x_1"]**2+ np.sin(my_inputs["x_2"]))
-        expression, info = model(target=y, **my_inputs)
-
-        my_vars = ",".join([key for key in my_inputs.keys()])
-        sp_expression = sp.sympify(expression)
-        fn_expression = sp.lambdify(my_vars, expression)
-        
-        _ = fn_expression(**my_inputs)
-
-        self.assertEqual(str, type(expression))
-        self.assertTrue(True)
 
 if __name__ == "__main__": #pragma: no cover
