+++ symr_tests/test_all.py
@@ -18,9 +18,11 @@ from symr_tests.symr.test_fake_sr import TestPolySR,\
         TestRandomSR,\
         TestLossFunction 
 
-from symr_tests.symr.test_wrappers import TestBaseWrapper,\
-        TestNSRTSWrapper,\
-        TestSymGPTWrapper
+from symr_tests.symr.test_wrappers import TestBaseWrapper
+from symr_tests.symr.test_nsrts_wrapper import TestNSRTSWrapper
+from symr_tests.symr.test_symgpt_wrapper import TestSymGPTWrapper
+from symr_tests.symr.test_symformer_wrapper import TestSymformerWrapper
+
 from symr_tests.symr.test_benchmark import TestEval,\
         TestBenchmark
